separate streams into separate files
