+++ command/dsop_bigbang.py
@@ -1,6 +1,7 @@
 from re import template
 from sqlite3 import Time
-from util.streams import *
+from util.streams import Stream
+from util.bb_streams import BigBang_Stream
 from appsettings import AppSettings
 import pathlib
 import util
